Splitting tests up
