+++ src/test/kotlin/com/nurflugel/hocon/PropertiesToConfSpec.kt
@@ -6,52 +6,6 @@ import com.nurflugel.hocon.parsers.PropertiesToConfParser.Companion.convertPrope
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.StringSpec
 
-/*
-
-Test cases:
-
-
- *
- * cors = [
- *    "some url"
- *    "another URL"
- *    "3 times a charm"
- *  ]
- * 
- * 
- * //comments
- * cors = [
- *    "some url"
- *    "another URL"
- *    "3 times a charm"
- *  ]
- * 
- * 
- * alpha.omega=false
- * // A comment before some code
- * alpha.beta.gamma=7
- * alpha.beta.delta=5
- * 
- * alpha.omega=false
- * // A comment before some code
- * // 2 comments before some code
- * alpha.beta.gamma=7
- * alpha.beta.delta=5
- * 
- * 
- * 
- * aaaa {
- * // A comment before some code
- *    bbbb {
- *         abab= 5
- *         // A comment before some code
- *         cccc = "text"
- *         dddd = true
- *      }
- *  }
- *  
- *  also - auto quote wrapping
- */
 class PropertiesToConfSpec : StringSpec(
   {
 
@@ -68,7 +22,6 @@ class PropertiesToConfSpec : StringSpec(
           }
 """)
     }
-
 
     "properties format simple keys map 2".config(enabled = ALL_TESTS_ENABLED) {
       val lines = getListFromString("""
@@ -87,7 +40,6 @@ class PropertiesToConfSpec : StringSpec(
           }
 """)
     }
-
 
     "properties format simple keys map 3".config(enabled = ALL_TESTS_ENABLED) {
       val lines = getListFromString("""
@@ -123,15 +75,12 @@ class PropertiesToConfSpec : StringSpec(
 """)
     }
 
-
-
     "properties format simple keys".config(enabled = ALL_TESTS_ENABLED) {
       val lines = getListFromString("""
         one="kkkk"
         two.three.four = 5
         """)
       val confLines = convertPropertiesToConf(lines)
-
 
       confLines shouldBe getListFromString("""
           one = "kkkk"
@@ -142,7 +91,6 @@ class PropertiesToConfSpec : StringSpec(
           }
 """)
     }
-
 
     "don't lose the includes in map formatter".config(enabled = ALL_TESTS_ENABLED) {
       val lines = getListFromString("""
@@ -173,100 +121,6 @@ class PropertiesToConfSpec : StringSpec(
       val propertyLines = convertConfToProperties(lines)
       val confLines = convertPropertiesToConf(propertyLines)
       confLines shouldBe getListFromString("aaaa.bbb.cccc.dddd = true")
-    }
-
-    "read in a single line list".config(enabled = ALL_TESTS_ENABLED) {
-      val propertyLines = getListFromString("""
-        aa="ff"
-        cors = ["123","456","789" ]
-        dd=false
-      """)
-      val confLines = convertPropertiesToConf(propertyLines)
-      confLines.forEach { println(it) }
-      confLines shouldBe getListFromString("""
-        aa = "ff"
-        cors = [
-          "123",
-          "456",
-          "789"
-        ]
-        dd = false
-      """.trimIndent())
-    }
-
-    /** I figure you should be all properties or not, but lists may be vertical, so we gotta deal with them */
-    "read in a multi-line list all lines separate".config(enabled = ALL_TESTS_ENABLED) {
-      val propertyLines = getListFromString("""
-        cors = [
-          "123",
-          "456",
-          "789"
-        ]
-      """)
-      val confLines = convertPropertiesToConf(propertyLines)
-      confLines shouldBe getListFromString("""
-        cors = [
-          "123",
-          "456",
-          "789"
-        ]
-      """.trimIndent())
-    }
-
-    "read in a multi-line list [ same lines".config(enabled = ALL_TESTS_ENABLED) {
-      val propertyLines = getListFromString("""
-        cors = [ "123",
-          "456",
-          "789"
-        ]
-      """)
-      val confLines = convertPropertiesToConf(propertyLines)
-      confLines shouldBe getListFromString("""
-        cors = [
-          "123",
-          "456",
-          "789"
-        ]
-      """.trimIndent())
-    }
-
-    "read in a multi-line list ] same lines".config(enabled = ALL_TESTS_ENABLED) {
-      val propertyLines = getListFromString("""
-        cors = [
-          "123",
-          "456",
-          "789" ]
-      """)
-      val confLines = convertPropertiesToConf(propertyLines)
-      confLines shouldBe getListFromString("""
-        cors = [
-          "123",
-          "456",
-          "789"
-        ]
-      """.trimIndent())
-    }
-
-
-    "read in a multi-line list with other keys ] same lines".config(enabled = ALL_TESTS_ENABLED) {
-      val propertyLines = getListFromString("""
-        abc = 678
-        cors = [
-          "123",
-          "456",
-          "789" ]
-        def=999
-      """)
-      val confLines = convertPropertiesToConf(propertyLines)
-      confLines shouldBe getListFromString("""
-        abc = 678
-        cors = [
-          "123",
-          "456",
-          "789"
-        ]
-        def = 999
-      """.trimIndent())
     }
 
   }) {
